demos/painter_path_test: make viewport coefficients equal
When stroking with pixel width, BAD things happen if the coefficient
going from normalized device coordinate to window coordinates are
not the same. This happens exactly when the glViewport() call has
that the width and height are not equal. To support pixel width
stroking we pass to glViewport a square that contains the window
and adjust the 2D projection matrix appropriately.