fastuidraw/stroking: Better pixel-width stroking in anisotropic case
It seems to never end. For StrokedPath, the normal for an edge is
now just the rotation of different vector. In addition, an edge
point has additional data for pixel width stroking to get the
"original" point of the data (Doxytags need update though).

For stroke shader in gl_backend/shaders/painter, massive
correction to fastuidraw_stroke_vert_align_normal_to_screen()
to actually give a vector so that when projected is perpindicular
to tangent vector. This made the quads coming from edge drawing
to now be proper rectangles for sheared-rotate-sheared transformations
tested by painter-path-test of demos. In addition rounded caps for
same tests are always discs now too. However, rounded joins are
still quite ugly (we need to figure "something" out to realize
the cos and sin data relative to the normals and tangents, but
not clear what since there are two frames(!).

In addition, have not yet tried out full blown 3D transformations
on the pixel-width stroking.