fastuidraw/glsl, painter: add first stroke interval start field to PainterDashedStrokeParams
The new field is a negative value representing when the first interval
(be it draw or space) starts. This value is either zero or negative.
It is negative if the first DashPatternElement has positive draw
length and the last has zero skip length or if the first has zero
draw length and the last has non-zero space length. The shaders to
compute the dash coverage uses this as the start of the first interval
instead of zero; doing so kills the 1-pixel missing content when drawing
anti-aliased closed dashing IF the dash pattern is such that the
start interval value is non-zero.