fastuidraw: preparation work for cap-joins for dashed stroking.
Skeleton code for drawing caps correctly on dashed stroking near
the ending of a join. The method is that the PainterAttributeData
will provide attributes for the cap-joins which are then
copied to a buffer in Painter work room which is then modified
by the DashEvaluator of the PainterDashedStrokeShaderSet.

We stil need to implement the creation of the cap-join data
in StrokedPath and the method adjust_cap_joins() for the
DashEvaluator of PainterDashedStrokeParams.