fastuidraw/glsl: Use distance to edge end to influence cap drawing
Compute the distance to the end of an edge and use that value to
cull those boundary points of the dash pattern past the end of
the edge when doing dashed stroking with caps.

We still need to cull on the start of the edge (i.e boundary
values before the start of the edge are also ignored). Also
TODO!

There is still an issue when on a sub-edge if such a boundary
is past the sub-edge and the angle between the neighboring sub-edge
is too large, then the cap's continuation is drawn with different
alignment. Solving this is on a TODO!

In addition, drawing caps at the start and end of a contour for
dashed stroking is not implemented yet either.