fastuidraw/glsl/shaders/painter: extend and shrink sub-edges for dashed stroking.
New code to extent or shrink a sub-edge for dashed stroking. The main
issue it fixes is when a cap for a dash boundary stretches over multiple
sub-edges and gets its shape distorted. This commit does the following:
  - sub-edges that are contained in a single dash interval are shrunk
    to the dash interval. If the dash interval is a skip-interval,
    then the sub-edge is collapsed by setting on_boundary to 0.
  - sub-edges that have a dash boundary within them, i.e. span
    multiple dash-intervals are expanded to include caps (if
    sub_shader_id indicates drawing with caps) and are also
    shrunk as well to not go past the cap (or if no caps are
    drawn, then to the dash intervals)

We have a rendering issue where if the a sub-edge extension
has its legnth extended to hit another dash boundary. then
the coverage comptation is wrong for rounded-caps.