fastuidraw/path: clean up tessellation interface of Path
Remove ability to set tess parameters for default tessellation.
Change the name of tessellation_lod(float) to tessellation(float).
If a non-positive value is passed, then return the default tessellation
which is made using the default values of TessellationParams. That
value is stored as the first element in the private array of
references to TessellatedPath objects.