fastidraw/painter/filled_path: more improvements on accuracy
GLU appears to be sometimes dropping an edge or merging two
points together that should NOT be merged. This can be seen
in the path file demo_data/paths/insane_path.txt where the
half circle arc on the left when it cuts through the path
below.

This commit makes it "harder" for the error to happen, but
it still does. If we did NOT apply a transformation to the
points, then the issue does not manifest.