fastuidraw/painter/stroked_path: StrokedPathSubset tracking joins and caps (Part 1)
Add to StrokedPathSubset tracking of what joins and caps
are inside of it along with what chunks would contain all
of the joins and caps inside of it. The chunk ordering
for the joins and caps is NOT the ordering currently
used for creation of joins and caps; instead the ordering
is dependent on the TessellatedPath being processed.