fastuidraw/util: rework DataBuffer
Rework DataBuffer to be a derived class for DataBufferBase
which is just a reference counterd holder for a
const_c_array<uint8_t> and c_array<uint8_t>. A DataBuffer
itself inherits from DataBufferBase and an object
that actually holds the data.

The eventual reason to have this is to support
DataBufferBase objects that source from mapped
memory.