fastuidraw/painter, packing, glsl, gl_backend: introduce concept Surface
Introduce new concept PainterBackend::Surface which represents
the surface to which to render content. Also implement a trivial
SurfaceGL which is backed by the default GL framebuffer.

Later patches will move data in PainterBackendGL out of
PainterBackendGL into SurfaceGL and also will change the
nature and interface of SurfaceGL in order to make possible
having the auxilary buffer backed by a render target instead
of an image so that one can use framebuffer fetch.

Change-Id: I6e33b367f7f343ebb640167df1b844c02fdb77a7