fastuidraw/painter/filled_path: fix tesselation error triggered by winding offset culling
It turns out GLU cannot handle correctly the situation if a single contour
has a loop within it. The fix is truly awful: we need to identify loops in
contours and break such loops into multiple contours.

The current implementation is not idea as will it take O(n^2) to identify
and remove loops, but there should be an implementation that is
O(n log n). However, given that these are contours after splitting
into subsets, the size is not horribly large (hopefully).

Change-Id: I4a35c562bd46d3fb14bd26449f88e3b60650c4d7