fastuidraw/painter, glsl: remove all anti-alias fuzz overdraw from stroking
Change the anti-alias fuzz to that it has no over-draw. This makes
shader-based anti-aliased stroking look mostly much better. In
exchange for removing the over on the inside of curves, we get that
on the corners where joins meet, a lack of anti-alias fuzz at essentially
one pixel (before the fuzz was drawn twice there often).

Change-Id: I5e3db5bf9083088f96be5dfdf5425232cc9aa88d