fastuidraw/painter/filled_path: make aa-edges occlude each other
To prevent overdraw of AA-fuzz in filling, make the AA-edges
occlude in Z, in the same way that stroking does Z-occlusion.
In addition, to make the number of chunks smaller, rather than
naming aa-fuzz chunks for each pair of winding numbers, instead
name it only for entire winding number components and rely
on that the solid fill will occlude the aa-edge fuzz.