fastuidraw/glsl/shaders/painter: arc-stroke dash shader numerical error fixing
When we do dashed arc-stroking, we rotate the radial vector back by the start
angle of the arc. For the points on the end of an arc we do this by first
rotating back by offset_vector and then rotate forward by the arc-angle.
However, this introduces numerical error enough that there are visual
issues when zooming in. To combat this, in the vertex shader, we apply
Gram-Shmidt to the rotatio matrix. This fixes the visual issue seen.