fastuidraw/painter: remove pixel gap at start join when anti-alias arc-dashed stroking without caps
The gap was caused by that the sample points along the boundary of the start
of the starting edge had a distance value of zero that when anti-aliased would
then give only partial coverage. The fix is to have the start distance a value
before 0 (we use negative the first draw length).

Change-Id: I1e38c1924fd84eb88f3c7a813e3a3f923b8f31d7