fastuidraw/path, tessellated_path: refactor tessellation interface
Change the label threshhold to max_distance since that is exactly
what it is: an estimate on the maximum distance between the path
and the tessellation of the path. In addition, have the meaning
of m_max_distance of <= 0.0f mean to choose a plain, default
tessellation (which means recurse the minimum required by the
edges for interpolator_generic and to have just one segment
for arcs). In addition tweak PathContour::arc and path_util_private
to handle correctly angles greater than PI.