fastuidraw/path: distance computation fix
The computation between a line-segment and point (used
to compute the upper-bound between a Bezier curve and
a line segment computation) actually computed the distance
between the line of a the given line segment and a point.
This patch corrects the computation by checking if the
canidate point of the line is in the line segment and if
it is, then use that; otherwise, use the max-value of the
distance between the point and the end-points of the line
segment.