fastuidraw/glsl/shader_source: avoid token pasting in macros.
GLES specification mandates that the token pasting is NOT supported
by the pre-processor. A number of GLES drivers support it anyways,
but our main target for GLES, Mesa, does not. This commit emulates
the macro FASTUIDRAW_LOCAL() by performing the token pasing itself.
However, the solution is not robust since it will utterly fall apart
if the argument to the macro is a macro itself that should be exapanded
or if the argument spans multiple lines. However, for most use-cases,
it is "good enough".