fastuidraw/path: rework and simplify arc choice
Rather than trying to force the mid-pt used to compute
the circle to be in the arc, instead simply take the
smaller arc of the two choices. The motivation is that
if it was getting linearly tessellated, the segment
would be the segment connecting the end points which
is the shortest curve between them; thus taking the
shortest arc is likely a better guess-choice.