fastuidraw/path: Change distance to arc-computation.
Sighs. NEVER trush ideas from the middle of the night before fully checking them.
The idea that using the max of costrol points to give the an upper-bound between
a Bezier curve and an an arc is WRONG. It is WRONG because an arc is not convex;
contrast this with a line which is convex. An easy counter example is to take
any two points on a circle and that the line segment from them. Clearly, the
distance to the arc is non-zero, but the two extreme points have zero-distance.

So to fix this up, instead make clear that the interface is to give an approximate
upper bound and to compute that as the max of distance of several points of the
curve, taking more points as the degree increases; the points taken are always
mid-points of a parent region, so the evaluation will pre-create not only the
children but potentially further descendants.