fastuidraw/text/private/rect_atlas: make rejection test much faster
The tracking of max-x and max-y sizes has a massive negative
performance impact. Instead of tracking those values, we do
a fast rejection test of rejecting any rectangle which has
a dimension atleast as large as a previously reject rectangle.
This test might reject rectangles that could have been taken
(for example if the last rejected rectangle was just too long
but was very short) any not short rectangle will get rejected.
However, since we are expecting glyphs to be not so far from
squares (and small as well), this case likely will not happen.