fastuidraw/text: New glyph renderer, RestrictedRays
A new glyph renderer using ideas found in the paper
"GPU-Centered Font Rendering Directly from Glyph Outlines" by
Eric Lengyel. We take the core ideas of correctly counting
winding value and performing anti-aliasing from the paper.
From there, we have additions. The main addition is that
we can restrict what curves to test to an arbitary box
subject to we have a point within the box where we know
the winding number. We implement a box hierarchy and
place four sample points within each leaf. The shader
then computes the winding "difference" between the
fragment position and the winding value at the position.
In addition, we enhance the anti-aliasing algorithm
against edges as well.