fastuidraw/tessellated_path, path: rework linearization.
Instead of having TessellatedPath store its own Path, have a
linearization() function to return the TessellatedPath with arc
converted to line-segments for a specified threshhold. In doing
so, a Path now only has a single tessellation() method that
returns an arc-tessellation. To get the linearization, then one
calls TessellatedPath::linearization(). The motivation for the
change is that this interface avoids accidental re-creation of
data that could result from TessellatedPath::path() and most
importantly, it removes having TessellatedPath have its own
Path which can be quite large from the arc-tessellation of a
starting Path.