fastuidraw/glsl: bug fix for drawing occluders
Rather than setting composite_shader to ~0u, set blend_shader to ~0u;
the reason is that when relying on single_src or dual_src blending,
the composite shader MUST run to correctly draw black. In contrast,
blend_shader is only active when one can read the color buffer value
from the frag shader and in those modes, the writes are effectively
skipped when occluder is getting drawn.