fastuidraw/text, int_path: fixes for approximating cubics via quadratics
The old code for approximating cubics by quadratics was unnecessarily
complicated and possibly wrong too. In addition, not applying a scaling
factor to the glyph data before conversion lost too much data.

This commit simplifies the approximating cubics via quadratics code
and applies a scaling factor to the glyph data for RestrictedRays
to get the render correct for a number of OTF out there.