fastuidraw/text/glyph_render_data_restricted_rays: rework coordinates.
Instead of biasing the coordinate values, we make translate the glyph
coordinate values so that all values are non-negative. With this change,
then we also need the vertex shader to pass the left/borrom coordinate
to the fragment shader since it is on longer implicitely zero.

Lastly, we also scale the incoming data so that the coordinate values
can fit within 16-bits.