fastuidraw/text/glyph_render_data_restricted_rays: massive rework.
Rework restricted rays glyph rendering:
1. Only a single sample point (at roughly the center of the box)
2. Change div-factor for sample point delta
3. Increase recursion limit (helps WildWood font rendering performance)
4. Compute coverage by computing L1-distance to nearest curve in -pixel-
   coordinates.

These changes give a massive rendering quality improvement.

The last step that remains is to instead of tracking a single
distance value to instead track the distance closest curve that
increments the winding number and the distance to the closest
curve that decrements the winding number. If the winding number
at the point is zero or if the fill rule is odd-even, the distance
is the minimum of the those two values. If the winding number is
-1, then the distance value is the distance to the closest curve
that increments. If the winding number is +1, then the distance
is the distance to the closest curve that decrements. Finally,
if the winding number has absolute value greater than or equal to
2, then we view the distance as outside of the pixel anyways.