fastuidraw/text/glyph_render_data_restricted_rays: track horizontal and vertical cancelling edges
Glyph 25561 of simsun.ttc has cancelling edges where two edges overlap
but go in opposite directions. This commit adds tracking of vertical
and horizontal edges that overlap and go in opposite directions. In
addition, it tracks partial overlap of seuch edges as well. These edges
are broken into sub-edges so that the overlap is exactly the same
and then we can mark such edges as not having impact and thus are
skipped from the the GPU data.