fastuidraw/glsl/shaders: optimize L1-distance computation
Change from having a tranformation to frag-coordinates from glyph coordinate
and performing the L1-distance computation in frag-coordinates to performing
the L1-distance computation in the coordinates of the transformation to the
reference point but we scale the values correctly. This change reduces the
computation some and -should- improve performance some.