fastuidraw/glsl/shaders/fastuidraw_restricted_rays: rework.
With some linear algebra, we can kill the anti-aliasing bleeding
that occured with the coverage computation that was combined
with the winding computation. The key ingredient is to make it
so that the composition of the window to glyph and glyph to local
coordinates is about unitary. We accomplish this by choosing the
top row vector of the glyph to local coordinates.