fastuidraw/text/glyph_selector: rework to avoid font duplication.
Sighs. GlyphSelector had a -terrible- bug behavior where when fonts
were added via generators, the generators would potentially be
used to create the same font multiple times, once for each
FontGroup it was added to.

This rework introduces the member class AbstractFont that
internally has both a FontGenerator and FontBase references
along with methods to fetch the font of the generator so
that a given generator only generates the font once.