fastuidraw/painter: define and implement begin_layer() and end_layer() interface
This large patch defines and implement fastuidraw::Painter's analgoue
of starting and ending a compositing/blending layer. In addition it
reworks begin() to take either a screen_orientation or a float3x3.
The former trivially wraps around the latter. For the latter, do
a clip_in_rect() passing the normalized device rect computed from
the viewport and then -concat- by the given transformation. Doing
this makes sure that the clipping region is tight as reasonably
possible.