fastuidraw/painter: fix clip changes after begin_layer()
Update the value of m_clip_rect to local coordinates on begin_layer().
Without this patch, calling begin_layer() followed by calling
clip_in_rect() would have clipping utterly incorrect and usually
result in NOTHING being draw. The second fix needed is to extend
the occluder drawing to [-3, 3] so that if clipping is set in a way
that generates occluders, the occluders will also block out writes
outside of the transparency areas region in the tranparency surface.