fastuidraw/painter/painter: add partial implementation to flush().
The method flush() will fail if the Painter is within a begin_layer()/end_layer()
pair (there is a way to make it make sense: just close up all layers but save
the surface before the layer closing). The bigger issue is that it will fail
if there are active occlusion interaction with the current state stack.
There is a way around this and placed as a TODO.