fastuidraw/painter: revamp PainterShaderData interfaces (part 1).
This first part revamps the PainterShaderData to remove utterly the DataBase
class and have PainterShaderData hold onto references to the resources
and the actual packed data.

The second part will remove from PainterPackedValue the copy
of the original data and have Painter and PainterPacker have
code to handle getting/creating the packed data.