fastuidraw/path: revamp PathContour interface
Rather than have the quite tricky code of tracking the predecessor interpolator_base
with a weak pointer and passing it on ctor, instead just pass the PathContour object
directly and have the interpolator_base ctor add itself to the PathContour's list
of interpolators. This also fixed a bug for open contours related to the fake
interpolator as well.

The only fly in the ointment is that the bounding box and flatness of the PathContour
cannot be updated when the interpolator is added because the functions determining
their values are pure virtual and can only be called after the object is completely
constructed.