fastuidraw/painter, glsl: add bounds checking to accessing deferred coverage buffer
Record to the PainterHeader the min and max corner within the deferred coverage
buffer of the allowed region to read. If a requested read goes outside of this
region, return 0.

The lack of bounds testing triggered a rendering issue in arc-stroke shading
sometimes because the area covered by the arcs are larger than the area occupied
by the actual stroking.