fastuidraw, demos: Limited support for Emscripten
Most of the work is of the form to decorate an amount of GL code
with #ifdef __EMSCRIPTEN__ and to remove any double quote (")
characters from shaders along with tweaking sdl_demo.cpp to
work with emscripten's interface. In addition, rather than adding
the emscripten target to the existing make-system, a new makefile
is added, Makefile.emscripten, which builds off the existing
Rules.mk of the make system. Minor tweaks are added to the build
system to make its build files not overlap with emscripten build
files. Lastly, the tutorials are NOT ported to emscripten.