fastuidraw/glsl, text: remove using patented technique
Remove using the patented techinque from Eric Lengyel. The technique
removed is where a 3-bit index is produced by comparing the three control
points of a quadratic bezier curve against 0 and from that index, using
a 2-bit valued truth table (enconded as a single 16-bit integer) to
decide is a solution to polynomial f(t) = 0.

To work around the patent but maintain numerical stabilty means
to examine the solutions from the quadratic formula,
t1 = (B - sqrt(B * B - AC)) / A, t2 = (B + sqrt(B * B - AC)) / A
against the condition 0 <= t1, t2 <=1. A hammer of algebra when
A > 0 yields the conditions:
  - use t1 if (q1 >= q2 && q1 >= 0.0) && (q3 > q2 || q3 < 0.0)
  - use t2 if (q3 > q2 && q3 > 0.0) && (q1 >= q2 || q1 <= 0.0)
where [q1, q2, q3] is the quadratic bezier curve. Note that
the comparisons involving the control point are against the
end points of the curve and not against 0.

The other claim of the patent of computing the anti-aliasing
coverage we never used because we tracked two distance values:
the "distance" to the nearest curves than increments the winding
and the "distance" to the nearest curves that decrement the winding.
The distance is almost the L1-distance, except that we omit checking
the end points of the curves. Using those two distance values
allows us to support both the non-zero and odd-even fill rules
simutaneously.

Comments:
  - for determining if a root should be used, the technique in the
    patent is faster
  - for the banded rays, the coverage technique in the patent together
    with tracking the nearest curve (not sorted by increment/decrement)
    gives better anti-aliasing under shearing, but does bleed. Using
    the distance-pair method in banded, looses some anti-aliasing
    under shearing. However, using the technique in the patent for
    anti-aliasing induces bleeding under shearing.
  - the restricted rays methods never used the anti-aliasing as in the
    parent and always used the tracking nearest curves; it gives great
    anti-aliasing because the fragment shader guarantess that the
    two axis used are perpindicular in SCREEN space.